% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_doe_workbench}
\alias{run_doe_workbench}
\alias{run_app}
\title{Launch the DoE Workbench Shiny App}
\usage{
run_doe_workbench(port = 4564L, launch.browser = TRUE, ...)

run_app(port = 4564L, launch.browser = TRUE, ...)
}
\arguments{
\item{port}{Port to run on (default: 4564)}

\item{launch.browser}{Whether to open a browser window (default: TRUE)}

\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\description{
Launch the DoE Workbench Shiny App
}
